backend/requirements.txt
flask
flask-cors
openai
beautifulsoup4
requests
pypdf
python-dotenv
werkzeug
weasyprint
markdown
